+++ tests/PhpPowerpoint/Tests/Writer/PowerPoint2007/SlideTest.php
@@ -168,7 +168,7 @@ class SlideTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
     }
     
-    public function testFillGradientLinear()
+    public function testFillGradientLinearTable()
     {
         $expected1 = 'E06B20';
         $expected2 = strrev($expected1);
@@ -193,32 +193,82 @@ class SlideTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expected2, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
     }
     
-    public function testFillGradientPath()
+    /**
+     * @link : https://github.com/PHPOffice/PHPPowerPoint/issues/61
+     */
+    public function testFillGradientLinearRichText()
+    {
+    	$expected1 = 'E06B20';
+    	$expected2 = strrev($expected1);
+    
+    	$phpPowerPoint = new PhpPowerpoint();
+    	$oSlide = $phpPowerPoint->getActiveSlide();
+    	$oShape = $oSlide->createRichTextShape();
+    	$oShape->setHeight(200)->setWidth(600)->setOffsetX(150)->setOffsetY(300);
+    	$oFill = $oShape->getFill();
+    	$oFill->setFillType(Fill::FILL_GRADIENT_LINEAR)->setStartColor(new Color('FF'.$expected1))->setEndColor(new Color('FF'.$expected2));
+    
+    	$pres = TestHelperDOCX::getDocument($phpPowerPoint, 'PowerPoint2007');
+    
+    	$element = '/p:sld/p:cSld/p:spTree/p:sp/p:spPr/a:gradFill/a:gsLst/a:gs[@pos="0"]/a:srgbClr';
+    	$this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
+    	$this->assertEquals($expected1, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
+    	$element = '/p:sld/p:cSld/p:spTree/p:sp/p:spPr/a:gradFill/a:gsLst/a:gs[@pos="100000"]/a:srgbClr';
+    	$this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
+    	$this->assertEquals($expected2, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
+    }
+    
+    public function testFillGradientPathTable()
+    {
+    	$expected1 = 'E06B20';
+    	$expected2 = strrev($expected1);
+    
+    	$phpPowerPoint = new PhpPowerpoint();
+    	$oSlide = $phpPowerPoint->getActiveSlide();
+    	$oShape = $oSlide->createTableShape(1);
+    	$oShape->setHeight(200)->setWidth(600)->setOffsetX(150)->setOffsetY(300);
+    	$oRow = $oShape->createRow();
+    	$oCell = $oRow->getCell();
+    	$oCell->createTextRun('R1C1');
+    	$oFill = $oCell->getFill();
+    	$oFill->setFillType(Fill::FILL_GRADIENT_PATH)->setStartColor(new Color('FF'.$expected1))->setEndColor(new Color('FF'.$expected2));
+    
+    	$pres = TestHelperDOCX::getDocument($phpPowerPoint, 'PowerPoint2007');
+    
+    	$element = '/p:sld/p:cSld/p:spTree/p:graphicFrame/a:graphic/a:graphicData/a:tbl/a:tr/a:tc/a:tcPr/a:gradFill/a:gsLst/a:gs[@pos="0"]/a:srgbClr';
+    	$this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
+    	$this->assertEquals($expected1, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
+    	$element = '/p:sld/p:cSld/p:spTree/p:graphicFrame/a:graphic/a:graphicData/a:tbl/a:tr/a:tc/a:tcPr/a:gradFill/a:gsLst/a:gs[@pos="100000"]/a:srgbClr';
+    	$this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
+    	$this->assertEquals($expected2, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
+    }
+    
+    /**
+     * @link : https://github.com/PHPOffice/PHPPowerPoint/issues/61
+     */
+    public function testFillGradientPathText()
     {
         $expected1 = 'E06B20';
         $expected2 = strrev($expected1);
         
         $phpPowerPoint = new PhpPowerpoint();
         $oSlide = $phpPowerPoint->getActiveSlide();
-        $oShape = $oSlide->createTableShape(1);
+        $oShape = $oSlide->createRichTextShape();
         $oShape->setHeight(200)->setWidth(600)->setOffsetX(150)->setOffsetY(300);
-        $oRow = $oShape->createRow();
-        $oCell = $oRow->getCell();
-        $oCell->createTextRun('R1C1');
-        $oFill = $oCell->getFill();
+        $oFill = $oShape->getFill();
         $oFill->setFillType(Fill::FILL_GRADIENT_PATH)->setStartColor(new Color('FF'.$expected1))->setEndColor(new Color('FF'.$expected2));
         
         $pres = TestHelperDOCX::getDocument($phpPowerPoint, 'PowerPoint2007');
         
-        $element = '/p:sld/p:cSld/p:spTree/p:graphicFrame/a:graphic/a:graphicData/a:tbl/a:tr/a:tc/a:tcPr/a:gradFill/a:gsLst/a:gs[@pos="0"]/a:srgbClr';
+        $element = '/p:sld/p:cSld/p:spTree/p:sp/p:spPr/a:gradFill/a:gsLst/a:gs[@pos="0"]/a:srgbClr';
         $this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
         $this->assertEquals($expected1, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
-        $element = '/p:sld/p:cSld/p:spTree/p:graphicFrame/a:graphic/a:graphicData/a:tbl/a:tr/a:tc/a:tcPr/a:gradFill/a:gsLst/a:gs[@pos="100000"]/a:srgbClr';
+        $element = '/p:sld/p:cSld/p:spTree/p:sp/p:spPr/a:gradFill/a:gsLst/a:gs[@pos="100000"]/a:srgbClr';
         $this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
         $this->assertEquals($expected2, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
     }
     
-    public function testFillPattern()
+    public function testFillPatternTable()
     {
         $expected1 = 'E06B20';
         $expected2 = strrev($expected1);
@@ -243,7 +293,7 @@ class SlideTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expected2, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
     }
     
-    public function testFillSolid()
+    public function testFillSolidTable()
     {
         $expected = 'E06B20';
     
@@ -262,6 +312,27 @@ class SlideTest extends \PHPUnit_Framework_TestCase
         $element = '/p:sld/p:cSld/p:spTree/p:graphicFrame/a:graphic/a:graphicData/a:tbl/a:tr/a:tc/a:tcPr/a:solidFill/a:srgbClr';
         $this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
         $this->assertEquals($expected, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
+    }
+
+    /**
+     * @link : https://github.com/PHPOffice/PHPPowerPoint/issues/61
+     */
+    public function testFillSolidText()
+    {
+    	$expected = 'E06B20';
+    
+    	$phpPowerPoint = new PhpPowerpoint();
+    	$oSlide = $phpPowerPoint->getActiveSlide();
+    	$oShape = $oSlide->createRichTextShape();
+    	$oShape->setHeight(200)->setWidth(600)->setOffsetX(150)->setOffsetY(300);
+    	$oFill = $oShape->getFill();
+    	$oFill->setFillType(Fill::FILL_SOLID)->setStartColor(new Color('FF'.$expected));
+    
+    	$pres = TestHelperDOCX::getDocument($phpPowerPoint, 'PowerPoint2007');
+    
+    	$element = '/p:sld/p:cSld/p:spTree/p:sp/p:spPr/a:solidFill/a:srgbClr';
+    	$this->assertTrue($pres->elementExists($element, 'ppt/slides/slide1.xml'));
+    	$this->assertEquals($expected, $pres->getElementAttribute($element, 'val', 'ppt/slides/slide1.xml'));
     }
     
     public function testHyperlink()